Fix: control page를 (withAuth) 폴더 내부로 변경
